Added more comments and split repository
