+++ refactor-me/Repositories/ProductsRepository.cs
@@ -79,73 +79,5 @@ namespace refactor_me.Repositories
             _productDataPool.CheckIn(productData);
             return result.ToProduct();
         }
-
-        public ProductOptions GetAllProductOptions(Guid productId)
-        {
-            var productData = _productDataPool.CheckOut();
-            var result = productData.ProductOptions.Where(item => item.ProductId == productId).ToList();
-            _productDataPool.CheckIn(productData);
-            return new ProductOptions(result.Select(item => item.ToProductOption()).ToList());
-        }
-
-        public ProductOption GetProductOption(Guid productId, Guid id)
-        {
-            var productData = _productDataPool.CheckOut();
-            var result = productData.ProductOptions.Where(item => item.ProductId == productId && item.Id == id).FirstOrDefault();
-            _productDataPool.CheckIn(productData);
-            return result?.ToProductOption();
-        }
-
-        public ProductOption CreateProductOption(Guid productId, ProductOption option)
-        {
-            var productData = _productDataPool.CheckOut();
-            var result = productData.ProductOptions.Add(new ProductOptionTable() { Id = Guid.NewGuid(), ProductId = productId, Name = option.Name, Description = option.Description });
-            productData.SaveChanges();
-            _productDataPool.CheckIn(productData);
-            return result.ToProductOption();
-        }
-
-        public ProductOption UpdateProductOption(Guid productId, Guid id, ProductOption option)
-        {
-            var productData = _productDataPool.CheckOut();
-            var result = productData.ProductOptions.Where(item => item.Id == id && item.ProductId == productId).FirstOrDefault();
-            if (result == null)
-            {
-                return null;
-            }
-            result.Name = option.Name;
-            result.Description = option.Description;
-            productData.SaveChanges();
-            _productDataPool.CheckIn(productData);
-            return result.ToProductOption();
-        }
-
-        public void DeleteProductOption(Guid productId, Guid id)
-        {
-            var productData = _productDataPool.CheckOut();
-            var toDelete = productData.ProductOptions.Where(item => item.Id == id && item.ProductId == productId).FirstOrDefault();
-            if (toDelete == null)
-            {
-                _productDataPool.CheckIn(productData);
-                return;
-            }
-            productData.ProductOptions.Remove(toDelete);
-            productData.SaveChanges();
-            _productDataPool.CheckIn(productData);
-        }
-
-        public void DeleteAllProductOptionsForProduct(Guid productId)
-        {
-            var productData = _productDataPool.CheckOut();
-            var toDelete = productData.ProductOptions.Where(item => item.ProductId == productId).ToList();
-            if (toDelete.Count == 0)
-            {
-                _productDataPool.CheckIn(productData);
-                return;
-            }
-            productData.ProductOptions.RemoveRange(toDelete);
-            productData.SaveChanges();
-            _productDataPool.CheckIn(productData);
-        }
     }
 }
