+++ refactor-me/refactor-me.csproj
@@ -137,7 +137,9 @@
     <Compile Include="Models\ProductOptions.cs" />
     <Compile Include="Models\Products.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Repositories\IProductOptionsRepository.cs" />
     <Compile Include="Repositories\IProductsRepository.cs" />
+    <Compile Include="Repositories\ProductOptionsRepository.cs" />
     <Compile Include="Repositories\ProductsRepository.cs" />
     <Compile Include="Services\IProductsService.cs" />
     <Compile Include="Services\ProductsService.cs" />
